Split ActionWait to it's own file.
